+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 * AES 128bit encryption to include unnecessary field in encryption dictionary to
   work around buggy PDF libraries
+* [HexaPDF::Layout::Style::LinkLayer] to correctly process the border color
 
 
 ## 0.32.0 - 2023-03-08